% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruvbox.R
\docType{data}
\name{gruvbox}
\alias{gruvbox}
\title{Function to access colors from a colorscheme}
\format{
An object of class \code{list} of length 2.
}
\usage{
gruvbox
}
\arguments{
\item{...}{Colors to extract.}
}
\value{
A named color hex code.
}
\description{
These functions are named after color schemes and are used to extract color 
hexcodes from a color scheme by the name of the color.
}
\examples{
gruvbox$dark("blue", "lightblue") # access colors
with(mtcars, plot(x = mpg, y = wt, col = dracula("pink"))) # use them inside a function
gruvbox$light("blue")
gruvbox$dark("blue")
monokai("blue")
dracula("blue")
}
\keyword{datasets}
