% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gruvbox.R
\name{gruvbox_pal}
\alias{gruvbox_pal}
\title{Function to access palettes}
\usage{
gruvbox_pal(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of the color palette.}

\item{reverse}{Logical; reverse the order of colors in the palette.}

\item{...}{Additional arguments to pass to colorRampPalette().}
}
\value{
A color palette.
}
\description{
Used to access palettes & use them with ggplot objects.
}
\examples{
dracula_pal("light")
dracula_pal("dark")
gruvbox_pal("dark")
monokai_pal("dark")
}
