% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_factories.R
\name{make_color_scale}
\alias{make_color_scale}
\title{Create color scale from palette}
\usage{
make_color_scale(.palette_fun, .name_pref)
}
\arguments{
\item{.palette_fun}{Function to create palettes, made with make_palette().}

\item{.name_pref}{Name prefix for color scale (invisible, but needed for ggplot2).}
}
\value{
A color scale function for use with ggplot objects.
}
\description{
Function factory to create color scales to use with a ggplot object from
custom palettes. Refer to the source code of ggschemes for example usage.
}
