% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_factories.R
\name{make_colorscheme}
\alias{make_colorscheme}
\title{Create color scheme}
\usage{
make_colorscheme(.colorscheme)
}
\arguments{
\item{.colorscheme}{Named vector with the color hexcodes (see examples below).}
}
\value{
A function to create color schemes.
}
\description{
Function factory to create a function that accesses color codes from a named 
vector. Used to make colors from a color scheme easily accessible by name without
conflicting with global color names.
}
\examples{
my_colors <- c(
  `black` = "#282a36",
  `white` = "#f8f8f2",
  `grey` = "#6272a4"
)

my_color <- make_colorscheme(my_colors)
my_color("black")
}
