% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_factories.R
\name{make_fill_scale}
\alias{make_fill_scale}
\title{Create fill scale from palette}
\usage{
make_fill_scale(.palette_fun, .name_pref)
}
\arguments{
\item{.palette_fun}{Function to create palettes, made with make_palette().}

\item{.name_pref}{Name prefix for color scale (invisible, but needed for ggplot2).}
}
\value{
A fill scale for use with ggplot objects.
}
\description{
Function factory to create fill scales to use with a ggplot object from
custom palettes. Refer to the source code of ggschemes for example usage.
}
