% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_theme_factories.R
\name{make_light_theme}
\alias{make_light_theme}
\title{Create a light ggplot theme from a template}
\usage{
make_light_theme(.template, .black, .grey, .lgrey, .white, ...)
}
\arguments{
\item{.template}{Function template to use.}

\item{.black}{Black color code.}

\item{.grey}{Grey color code.}

\item{.lgrey}{Light grey color code.}

\item{.white}{White color code.}

\item{...}{Additional arguments to pass to theme().}
}
\description{
This function factory is used to create a light ggplot theme from a template
giving some colors which are used for the background, text, axes and grid lines.
}
