% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monokai.R
\name{monokai}
\alias{monokai}
\title{Function to access colors from a colorscheme}
\usage{
monokai(...)
}
\arguments{
\item{...}{Colors to extract.}
}
\value{
A named color hex code.
}
\description{
These functions are named after color schemes and are used to extract color 
hexcodes from a color scheme by the name of the color.
}
\examples{
gruvbox$dark("blue", "lightblue") # access colors
with(mtcars, plot(x = mpg, y = wt, col = dracula("pink"))) # use them inside a function
gruvbox$light("blue")
gruvbox$dark("blue")
monokai("blue")
dracula("blue")
}
