% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula.R
\name{scale_color_dracula}
\alias{scale_color_dracula}
\title{Color scales for ggschemes}
\usage{
scale_color_dracula(palette = "light", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette to use for constructing the color scale. Defaults to "light".}

\item{discrete}{Logical; true means discrete scale, false means continuous scale.}

\item{reverse}{Logical; reverse the order of the color scale?}
}
\description{
Color scales for ggschemes' color schemes and themes, to be used with ggplot
objects.
}
\examples{
mtcars |> 
  ggplot(aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  theme_dracula_dark() +
  scale_color_dracula(palette = "dark")
mtcars |> 
  ggplot(aes(wt, mpg, color = wt)) +
  geom_point() +
  theme_dracula_light() +
  scale_color_dracula(palette = "rank", discrete = FALSE, reverse = TRUE)
}
