% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula.R
\name{scale_fill_dracula}
\alias{scale_fill_dracula}
\title{Fill scales for ggschemes}
\usage{
scale_fill_dracula(palette = "light", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette to use for constructing the color scale. Defaults to "light".}

\item{discrete}{Logical; true means discrete scale, false means continuous scale.}

\item{reverse}{Logical; reverse the order of the color scale?}
}
\description{
Fill color scales for ggschemes' color schemes and themes, to be used with ggplot
objects.
}
\examples{
mtcars |> 
  dplyr::count(cyl) |> 
  ggplot(aes(factor(cyl), n, fill = factor(cyl))) +
  geom_col() +
  theme_monokai_dark() +
  scale_fill_monokai(palette = "dark")
}
