% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula.R
\name{theme_dracula_light}
\alias{theme_dracula_light}
\title{Plot themes from ggschemes}
\usage{
theme_dracula_light(...)
}
\arguments{
\item{...}{Additional arguments to pass to theme().}
}
\description{
These are the themes exported from ggschemes. The themes themselves simply 
apply the colors determined by the function factory to a theme template, so if
you want to modify the plot theme, you are likely better off modifying the
underlying template, or simply adding a theme()-call to your plot to override 
the settings made by these themes. Best used in conjunction with the matching
color scale.
}
\examples{
mtcars |> 
  ggplot(aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  theme_dracula_light() +
  scale_color_dracula()
}
