% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_theme_templates.R
\name{theme_template}
\alias{theme_template}
\title{Template for creating ggplot themes}
\usage{
theme_template(...)
}
\arguments{
\item{...}{Additional arguments to pass to theme().}
}
\description{
This is a theme template used for creating themes within ggschemes. It is similar
to theme_minimal().
}
