% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_theme_templates.R
\name{theme_template_horizontal}
\alias{theme_template_horizontal}
\title{Template for creating ggplot themes}
\usage{
theme_template_horizontal(...)
}
\arguments{
\item{...}{Additional arguments to pass to theme().}
}
\description{
This is a theme template used for creating themes within ggschemes. This template
emphasizes the horizontal gridlines.
}
